fix: array map안에서 훅스를 사용해서 생긴 에러 해결
